check for output during discovar run

+++ runDiscovarVar.sh
@@ -18,11 +18,21 @@ echo "LANE " $LANE
 echo "DATA " $DATA
 echo "REGION " $REGION
 
+ls ${NAME}.final.variant.filtered.vcf
+rc=$?;
+if [[ $rc == 0 ]];
+then
+    echo "DISCOVAR OUTPUT ALREADY PRESENT, NOT RUNNING";
+    exit 0;
+fi
+
+
 echo "LINK/INDEX BAM..."
 BAMFILE=`ls ${DATA}/${LANE}/${SET}/*bam`
 
 ln -s $BAMFILE ${NAME}.bam
 samtools index ${NAME}.bam
+
 
 echo "DISCOVAR"
 Discovar READS=${NAME}.bam \
@@ -30,12 +40,33 @@ Discovar READS=${NAME}.bam \
 	 OUT_HEAD=${NAME} \
 	 TMP=./tmp \
 	 REFERENCE=${WORK}/refs/PlasmoDB-24_Pfalciparum3D7_Genome.fasta
+rc=$?;
+if [[ $rc != 0 ]];
+then
+    echo "DISCOVAR ERROR";
+    exit $rc;
+fi
+
+ls ${NAME}.final.variant.filtered.vcf
+rc=$?;
+if [[ $rc != 0 ]];
+then
+    echo "DISCOVAR FINISHED WITHOUT OUTPUT";
+    exit $rc;
+fi
+
 
 echo "MAKE OUTPUTS"
 dot -Tpng -o ${NAME}.final.png ${NAME}.final.dot
 perl -i -pe "s/${SET}$/${NAME}/gi" ${NAME}.final.variant.filtered.vcf
 bgzip ${NAME}.final.variant.filtered.vcf
 tabix ${NAME}.final.variant.filtered.vcf.gz
+
+if [[ $rc != 0 ]];
+then
+    echo "TABIX/BGZIP ERROR";
+    exit $rc;
+fi
 
 echo "CLEANUP"
 rm -r tmp
